+++ doc/conf.py
@@ -50,6 +50,7 @@ extensions = [
     "sphinx.ext.viewcode",
     "sphinx.ext.intersphinx",
     "sphinx_copybutton",
+    "sphinx_gallery.gen_gallery",
     "m2r2"
 ]
 
@@ -108,16 +109,16 @@ html_sidebars = {
 # Xanadu theme options (see theme.conf for more information).
 html_theme_options = {
     "navbar_name": "Xanadu Sphinx Theme",
-    "navbar_logo_colour": "#2d7c7f",
+    "navbar_logo_colour": "#f48fb1",
     # "navbar_home_link": "https://pennylane.ai",
     "navbar_left_links": [
         {
-            "name": "Example1",
+            "name": "Blog",
             "href": "javascript:void(0);",
         },
         {
-            "name": "Blog",
-            "href": "javascript:void(0);",
+            "name": "Search",
+            "href": "search.html",
         },
         {
             "name": "Documentation",
@@ -141,5 +142,31 @@ html_theme_options = {
     "table_header_background_colour": "#f8bbd0",
     "text_accent_colour": "#f48fb1",
     "toc_marker_colour": "#f48fb1",
-    "toc_overview": True
+    "toc_overview": True,
+    "toc_global": True,
+    "toc_hover": False,
+    "relations": True
+}
+
+
+# -- Options for Sphinx Gallery --------------------------------------------
+
+sphinx_gallery_conf = {
+    # path to your example scripts
+    "examples_dirs": ["tutorials"],
+    # path where to save gallery generated examples
+    "gallery_dirs": ["gallery"],
+    # execute files that match the following filename pattern,
+    # and skip those that don't. If the following option is not provided,
+    # all example scripts in the 'examples_dirs' folder will be skiped.
+    "filename_pattern": r"tutorial",
+    "pypandoc": True,
+    # first notebook cell in generated Jupyter notebooks
+    "first_notebook_cell": (
+        "# This cell is added by sphinx-gallery\n"
+        "# It can be customized to whatever you like\n"
+        "%matplotlib inline"
+    ),
+    # thumbnail size
+    "thumbnail_size": (400, 400),
 }