+++ xanadu_sphinx_theme/directives/__init__.py
@@ -2,6 +2,7 @@
 This package implements custom Xanadu Sphinx directives.
 """
 
+from .bio import BioDirective
 from .community_card import CommunityCardDirective
 from .details import DetailsDirective
 from .gallery_item import GalleryItemDirective
@@ -11,6 +12,7 @@ from .title_card import TitleCardDirective
 from .youtube_video import YouTubeVideoDirective
 
 __all__ = [
+    "BioDirective",
     "CommunityCardDirective",
     "DetailsDirective",
     "GalleryItemDirective",